Updated SuperUserDashboard.tsx
